[Chore] Organize components booth folder
