+++ src/components/booth/pub/PubCarousel.jsx
@@ -5,11 +5,12 @@ import 'slick-carousel/slick/slick-theme.css';
 import styled from 'styled-components';
 import arrowKeyLeft from '@/assets/webps/booth/icon/arrowKeyLeft.webp';
 import arrowKeyRight from '@/assets/webps/booth/icon/arrowKeyRight.webp';
-import CarouselItem from '@/components/booth/CarouselItem';
+import CarouselItem from '@/components/booth/pub/CarouselItem';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
 import { PubList } from '@/constants/booth/pubList';
 import { useState, useEffect, useRef } from 'react';
+import { setCategoryFunc } from '@/utils/booth/setCategoryFunc';
 
 PubCarousel.propTypes = {
   menu: PropTypes.string,
@@ -24,47 +25,12 @@ export default function PubCarousel({ menu, click, likeData }) {
   const sliderRef = useRef(null);
 
   useEffect(() => {
-    const newCategory = setCategoryFunc(menu);
+    const newCategory = setCategoryFunc(menu, t);
     setCurrentCategory(newCategory);
     if (sliderRef.current) {
       sliderRef.current.slickGoTo(0);
     }
   }, [menu]);
-
-  const setCategoryFunc = (menu) => {
-    switch (menu) {
-      case t('booth.pub.menu.1'):
-        return '1';
-      case t('booth.pub.menu.2'):
-        return '2';
-      case t('booth.pub.menu.3'):
-        return '3';
-      case t('booth.pub.menu.4'):
-        return '4';
-      case t('booth.pub.menu.5'):
-        return '5';
-      case t('booth.pub.menu.6'):
-        return '6';
-      case t('booth.pub.menu.7'):
-        return '7';
-      case t('booth.pub.menu.8'):
-        return '8';
-      case t('booth.pub.menu.9'):
-        return '9';
-      case t('booth.pub.menu.10'):
-        return '9';
-      case t('booth.pub.menu.11'):
-        return '10';
-      case t('booth.pub.menu.12'):
-        return '11';
-      case t('booth.pub.menu.13'):
-        return '12';
-      case t('booth.pub.menu.14'):
-        return '13';
-      default:
-        return '1';
-    }
-  };
 
   const CustomArrow = ({ className, onClick, direction }) => {
     return (
