+++ src/components/booth/ranking/Kind111.jsx
