+++ src/components/booth/ranking/Kind113.jsx
