+++ src/components/booth/ranking/Kind122.jsx
