+++ src/components/booth/ranking/Kind123.jsx
