+++ src/i18n/resource/ko.json
@@ -377,14 +377,39 @@
     "location": "플리마켓 위치",
     "list": [
       {
-        "key": "ccJuice",
-        "name": "씨씨주스",
-        "intro": "신선한 과일로 만든 주스와 화채로 더운\n 날씨를 이겨낼 수 있는 곳"
+        "key": "hypeBoy",
+        "name": "홍익 Hype Boy 마켓",
+        "intro": "슈프림, 팔라스, 스투시, 나이키 등 다양한 스트릿\n 브랜드 의류 저렴하게 가져가세요!"
       },
       {
         "key": "olmak",
         "name": "올막 (Olmak)",
         "intro": "자연에서 영감을 받은 도자와 유리 소재의\n 핸드메이드 주얼리와 소품을 팝니다"
+      },
+      {
+        "key": "aiesec",
+        "name": "AIESEC과 함께 MZ한 친환경 실천 어때?",
+        "intro": "다회용 텀블러 & 키링으로 변신 가능한 손뜨개\n 티코스터"
+      },
+      {
+        "key": "sangsu",
+        "name": "상수좌판",
+        "intro": "국내외 각종 중고 LP와 CD 등 음반 판매"
+      },
+      {
+        "key": "ccJuice",
+        "name": "씨씨주스",
+        "intro": "신선한 과일로 만든 주스와 화채로 더운\n 날씨를 이겨낼 수 있는 곳"
+      },
+      {
+        "key": "sajuraplz",
+        "name": "sajuraplz",
+        "intro": "내가 만든 도자기 너를 위해 구웠지"
+      },
+      {
+        "key": "kawaii",
+        "name": "겨운거 ^..^",
+        "intro": "손으로 만든 특별한 뜨개 작품으로 귀여움과\n 애정을 전합니다"
       },
       {
         "key": "henna",
@@ -395,31 +420,6 @@
         "key": "modori",
         "name": "모도리",
         "intro": "뚝딱뚝딱 목수들의 아기자기한 소품 가게"
-      },
-      {
-        "key": "sajuraplz",
-        "name": "sajuraplz",
-        "intro": "내가 만든 도자기 너를 위해 구웠지"
-      },
-      {
-        "key": "hypeBoy",
-        "name": "홍익 Hype Boy 마켓",
-        "intro": "슈프림, 팔라스, 스투시, 나이키 등 다양한 스트릿\n 브랜드 의류 저렴하게 가져가세요!"
-      },
-      {
-        "key": "sangsu",
-        "name": "상수좌판",
-        "intro": "국내외 각종 중고 LP와 CD 등 음반 판매"
-      },
-      {
-        "key": "kawaii",
-        "name": "겨운거 ^..^",
-        "intro": "손으로 만든 특별한 뜨개 작품으로 귀여움과\n 애정을 전합니다"
-      },
-      {
-        "key": "aiesec",
-        "name": "AIESEC과 함께 MZ한 친환경 실천 어때?",
-        "intro": "다회용 텀블러 & 키링으로 변신 가능한 손뜨개\n 티코스터"
       }
     ],
     "detail": [
