+++ src/utils/booth/setCategoryFunc.js
@@ -1,0 +1,34 @@
+export const setCategoryFunc = (menu, t) => {
+  switch (menu) {
+    case t('booth.pub.menu.1'):
+      return '1';
+    case t('booth.pub.menu.2'):
+      return '2';
+    case t('booth.pub.menu.3'):
+      return '3';
+    case t('booth.pub.menu.4'):
+      return '4';
+    case t('booth.pub.menu.5'):
+      return '5';
+    case t('booth.pub.menu.6'):
+      return '6';
+    case t('booth.pub.menu.7'):
+      return '7';
+    case t('booth.pub.menu.8'):
+      return '8';
+    case t('booth.pub.menu.9'):
+      return '9';
+    case t('booth.pub.menu.10'):
+      return '9';
+    case t('booth.pub.menu.11'):
+      return '10';
+    case t('booth.pub.menu.12'):
+      return '11';
+    case t('booth.pub.menu.13'):
+      return '12';
+    case t('booth.pub.menu.14'):
+      return '13';
+    default:
+      return '1';
+  }
+};
